backend/pyproject.toml
[project]
name = "crawsla-backend"
version = "0.1.0"
description = "Crawsla backend — scraper + API for car listings"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0",
    "pg8000>=1.31",
    "python-dotenv>=1.0",
    "playwright>=1.48",
    "typer>=0.13",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "scraper*"]
